requirements.txt
gt-telem>=0.1.0
requests>=2.25.0
PyYAML>=5.4.0
